#123 sealed interface DataEntity and moved CustomDataEntity to entities package
