+++ core/src/main/java/contacts/core/entities/CustomData.kt
@@ -1,6 +1,7 @@
-package contacts.core.entities.custom
+package contacts.core.entities
 
-import contacts.core.entities.*
+// Note that this is declared here and not in the entities.custom package because DataEntity is sealed here.
+// If Kotlin increases the scope of "sealed" from package to module, then we can move it back =)
 
 /**
  * A custom [DataEntity].
