+++ core/src/main/java/contacts/core/entities/custom/AbstractCustomDataOperation.kt
@@ -2,6 +2,7 @@ package contacts.core.entities.custom
 
 import contacts.core.AbstractCustomDataField
 import contacts.core.entities.MimeType
+import contacts.core.entities.MutableCustomData
 import contacts.core.entities.operation.AbstractDataOperation
 
 /**
